Add support for negative amounts in billables (storno/cancellation lines)
This change enables negative values in both the amount and price_per_unit
fields of billables.csv, allowing users to create storno (reversal/
cancellation) lines that reduce invoice totals.

Changes:
- format.zig: Updated intThousands() and floatThousands() to handle
  negative numbers by detecting sign, working with absolute values,
  and prepending minus sign
- json.zig: Changed Offer.total and Invoice.total from ?usize to ?i64
- fj.zig: Updated billables calculation pipeline to use i64 throughout
  (grand_total, line_price, vat_amount, group_sum_map values)
- fj.zig: Changed price_per_unit parsing from parseUnsigned to parseInt
- ep_utils.zig: Updated Stats struct amount fields from usize to i64
- fj.zig: Updated billables.csv template documentation with examples
  of negative amounts and storno lines
- format.zig: Added comprehensive tests for negative integers and floats

Example usage in billables.csv:
  Storno by price, 1, pauschal, -1000, (cancellation)
  Storno by amount, -2.5, week, null, (reversed weeks)

The implementation is backward compatible - all existing positive-only
invoices continue to work correctly.

+++ src/web/ep_utils.zig
@@ -71,10 +71,10 @@ pub const Stats = struct {
     num_invoices_total: isize = 0,
     num_offers_open: isize = 0,
     num_offers_total: isize = 0,
-    invoiced_total_amount: usize = 0,
-    invoices_open_amount: usize = 0,
-    offers_pending_amount: usize = 0,
-    offers_accepted_amount: usize = 0,
+    invoiced_total_amount: i64 = 0,
+    invoices_open_amount: i64 = 0,
+    offers_pending_amount: i64 = 0,
+    offers_accepted_amount: i64 = 0,
 };
 
 pub fn show_404(arena: Allocator, context: *Context, r: zap.Request) !void {
